% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBigWig.R
\name{fetch_region}
\alias{fetch_region}
\title{Fetch the score-values from a BigWig file in the provided region.}
\usage{
fetch_region(path, chromosome, start, end)
}
\arguments{
\item{path}{A non-empty character string pointing to an existing file}

\item{start}{An integer value defining the start position (0-based half-open)}

\item{end}{An integer value defining the end position (0-based half-open). Must be larger than `start`}

\item{chromsome}{The name of the chromosome to fetch from}
}
\value{
A data.frame with either 2 or three columns containing either
 `Position` and `Score` or `Start`, `End`, `Score` respectively.
}
\description{
Fetch the score-values from a BigWig file in the provided region.
}
