% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBigWig.R
\name{fetch_region_means}
\alias{fetch_region_means}
\title{Calculate a `mean`-statistic on `bins` sub-intervals in the provided region.}
\usage{
fetch_region_means(
  path,
  chromosome,
  start,
  end,
  bins = 1,
  as_data_frame = FALSE
)
}
\arguments{
\item{path}{A non-empty character string pointing to an existing file}

\item{start}{An integer value defining the start position (0-based half-open)}

\item{end}{An integer value defining the end position (0-based half-open). Must be larger than `start`.}

\item{bins}{The number of sub-bins to generate.}

\item{as_data_frame}{Return a data.frame with precalculated start and ends of the intervals.}

\item{chromsome}{The name of the chromosome to fetch from}
}
\value{
A numeric vector of length `bins` or a `data.frame` with `bins` rows and the columns
`Start`, `End`, `Bin`, and `Score`.
}
\description{
Calculate a `mean`-statistic on `bins` sub-intervals in the provided region.
}
